Bind vertical profiles into time series objects
Source: R/bioRad.R

Binds profiles and list of profiles into a time series object. Also combines multiple time series objects of a single radar into one time series object.

bind(x, ...)

# S3 method for vp
bind(...)

# S3 method for vplist
bind(x, ...)

# S3 method for vpts
bind(..., attributes.from = 1)

Arguments

x
objects of class vp, vplist or vpts

...
object of class vp or vpts

attributes.from
which vpts object to copy attributes form (default: first)

Value

An object or a list of objects of class vpts, see summary.vpts for details The profiles in the input vpts objects will be sorted in time in the output object.

Methods (by class)

vp: bind vp objects into a vpts object. If vp of multiple radars are provided, a list is returned containing vpts time series objects for each radar.

vplist: bind vplist objects into a vpts object. If data of multiple radars is provided, a list is returned containing vpts time series objects for each radar.

vpts: bind multiple time series of vertical profiles (vpts objects) into a single vpts object. Requires the vpts objects to be from the same radar.

Examples

# NOT RUN {
# load the example vpts object
data(VPTS)
# split the vpts object into two separate time series,
# one containing profile 1-10, and a second combining profile 11-20:
vpts1=VPTS[1:10]
vpts2=VPTS[11:20]
# use bind to merge the two together:
vpts1and2=bind(vpts1,vpts2)
# verify that the binded objected now has 20 profiles, 10 from vpts1 and 10 from vpts2:
summary(vpts1and2)
# extract two profiles
vp1=VPTS[1]
vp1
vp2=VPTS[2]
vp2
# bind the profile objects back into a time series object:
bind(vp1,vp2)
# }