Bind vertical profiles (vp) into time series (vpts)

Binds vertical profiles (vp) into a vertical profile time series (vpts), sorted in time. Can also bind multiple vpts of a single radar into one vpts.

bind_into_vpts(x, ...)

# S3 method for vp
bind_into_vpts(...)

# S3 method for list
bind_into_vpts(x, ...)

# S3 method for vpts
bind_into_vpts(..., attributes_from = 1)

Arguments

x
A vp, vpts or a vector of these.

...
A vp, vpts or a vector of these.

attributes_from
integer. Which vpts to copy attributes from (default: first).

Value

A vpts for a single radar or a list of vpts for multiple radars. Input vp are sorted in time in the output vpts.

Methods (by class)

vp: Bind multiple vp into a vpts. If vp for multiple radars are provided, a list is returned containing a vpts for each radar.

list: Bind multiple vp objects into a vpts. If data for multiple radars is provided, a list is returned containing a vpts for each radar.

vpts: Bind multiple vpts into a single vpts. Requires the input vpts to be from the same radar.

Examples

# load example time series of vertical profiles:
data(example_vpts)

# split the vpts into two separate time series, one containing profile 1-10,
# and a second containing profile 11-20:
vpts1 <- example_vpts[1:10]
vpts2 <- example_vpts[11:20]

# use bind_into_vpts to bind the two together:
vpts1and2 <- bind_into_vpts(vpts1, vpts2)

# verify that the binded vpts now has 20 profiles, 10 from vpts1 and 10 from
# vpts2:
summary(vpts1and2)

#>                    Regular time series of vertical profiles (class vpts)
#> 
#>            radar:  KBGM 
#>       # profiles:  20 
#> time range (UTC):  2016-09-01 00:02:00 - 2016-09-01 01:18:00 
#>    time step (s):  240 

# extract two profiles:
vp1 <- example_vpts[1]
vp1

#>                Vertical profile (class vp)
#> 
#>        radar:  KBGM 
#>       source:  
#> nominal time:  2016-09-01 00:02:00 
#> generated by:   

vp2 <- example_vpts[2]
vp2

#>                Vertical profile (class vp)
#> 
#>        radar:  KBGM 
#>       source:  
#> nominal time:  2016-09-01 00:06:00 
#> generated by:   

# bind the two profiles back into a vpts:
bind_into_vpts(vp1, vp2)

#>                    Regular time series of vertical profiles (class vpts)
#> 
#>            radar:  KBGM 
#>       # profiles:  2 
#> time range (UTC):  2016-09-01 00:02:00 - 2016-09-01 00:06:00 
#>    time step (s):  240