Deprecated functions in package bioRad

The functions listed below are deprecated or renamed and will be defunct in the near future. When possible, alternative functions with similar functionality are mentioned.

basemap(...)

beamheight(range, elev, k = 4/3, lat = 35, re = 6378, rp = 6357)

beamwidth(range, angle = 1)

bind(x, ...)

checkDocker(...)

composite(...)

cmt(x, alt.min = 0, alt.max = Inf, alpha = NA, interval.max = Inf)

day(x, elev = -0.268)

dbz2eta(...)

download_vp(...)

elangle(...)

eta2dbz(...)

fetch(...)

getscan(...)

h5ODIMobject(...)

is.vplist(...)

mt(x, alt.min = 0, alt.max = Inf, alpha = NA, interval.max = Inf)

mtr(x, alt.min = 0, alt.max = Inf, alpha = NA)

night(x, ..., elev = -0.268)

ppi(...)

read.pvol(...)

readvp(...)

readvp.list(...)

readvp.table(...)

regularize(...)

retrieve_vp_paths(...)

rsl2odim(...)

sd_vvp(...)

sd_vvp(x) <- value

suntime(..., rise = TRUE)

updateDocker(...)

vintegrate(...)

vol2bird(...)

vpts(...)

SCAN(...)

VP(...)

VPTS(...)

basemap

Use download_basemap instead.

beamheight

Use beam_height instead.

beamwidth

Use beam_width instead.

bind

Use bind_into_vpts instead.

checkDocker

Use check_docker instead.

composite

Use composite_ppi instead.

cmt

cmt is deprecated as a separate function. Cumulative migration traffic is now included in the output of integrate_profile as column mt, which can be summed to get the cumulative migration traffic.

day

Use check_night instead.

dbz2eta

Use dbz_to_eta instead.

download_vp

Use download_vpfiles instead.

elangle

Use get_angles instead.

eta2dbz

Use eta_to_dbz instead.

fetch

Use get_quantity instead.

getscan

Use get_scan instead.

h5ODIMobject

Use get_odim_object_type instead.

is.vplist

The vplist object is deprecated. Use a regular list of vp objects (c(vp, vp)) instead.

mt

mt is deprecated as a separate function. Migration traffic is now included in the output of integrate_profile as column mt.

mtr

mtr is deprecated as a separate function. Migration traffic rate is now included in the output of integrate_profile as column mtr.

night

Use check_night instead.

ppi

Use project_as_ppi instead.

read.pvol

Use read_pvolfile instead.

readvp

Use read_vpfiles instead.

readvp.list

Use read_vpfiles instead.

readvp.table

Use read_vpts instead.

regularize

Use regularize_vpts instead.

retrieve_vp_paths

Use select_vpfiles instead.

rsl2odim

Use nexrad_to_odim instead.

sd_vvp

Use rvsd instead.

sd_vvp&lt;-

Use rvsd<- instead.

suntime

Use sunrise or sunset instead.

updateDocker

Use update_docker instead.

vintegrate

Use integrate_profile instead.

vol2bird

Use calculate_vp instead.

vpts

Use bind_into_vpts instead.

SCAN

Use example_scan instead.

VP

Use example_vp instead.

VPTS

Use example_vpts instead.