bioRad: Biological analysis and visualization of weather radar data

bioRad aims to provide standardized methods for extracting and reporting biological signals from weather radars. It includes functionality to inspect low‐level radar data, process these data into meaningful biological information on animal speeds and directions at different altitudes in the atmosphere, visualize these biological extractions, and calculate further summary statistics.

Details

To get started, see:

Dokter et al. (2018): a paper describing the package.

bioRad vignette: an introduction to bioRad's main functionalities.

Function reference: an overview of all bioRad functions.

See also