Cumulative migration traffic

Cumulative migration traffic is calculated as the cumulative sum of the migration traffic within each time step of a time series. Cumulative migration traffic gives the number of individuals that have passed per km perpendicular to the migratory direction at the position of the radar as a function oftime from the start of time series within the specified altitude band.

cmt(x, alt.min = 0, alt.max = Inf, alpha = NA)

Arguments

x
an object inhereting from class 'vpts'

alt.min
minimum altitude in m

alt.max
maximum altitude in m

alpha
(optional) migratory direction of interest in clockwise degrees from north, otherwise NA

Value

a numeric value equal to migration traffic in number of individuals / km

Examples

# NOT RUN {
# get the VPTS example dataset:
data(VPTS)
# print cumulative migration traffic to console:
cmt(VPTS)
# plot cumulative migration traffic:
plot(cmt(VPTS),type='l',xlab="time",ylab="CMT [birds/km]")
# }