Create a composite of multiple plan position indicators (ppi)

Combines multiple plan position indicators (ppi) into a single ppi. Can be used to make a composite of ppi's from multiple radars.

composite_ppi(x, param = "DBZH", dim = c(100, 100))

Arguments

x
A list of ppi.

param
Scan parameter to composite.

dim
integer. Vector with number of cells in each spatial dimension.

Value

A ppi.

Details

This function is a prototype, wait for the next bioRad release for more documentation and functionality.

The latitude/longitude of the returned ppi use the WGS84 datum.

Examples

# load the example polar scan:
data(example_scan)
# to be written ...