Download vertical profile files (vp) from the ENRAM data repository

Download a set of vp bird profiles from the ENRAM repository. These are stored within monthly available zip folders. This function downloads and unzips them at a user defined location. Check http://enram.github.io/data-repository/ for an overview of available data.

download_vpfiles(date_min, date_max, country, radar, directory = ".")

Arguments

date_min
ISO fomat date indicating the first date to download files from.

date_max
ISO fomat date indicating the last date to download files from

country
Char vector with two letter country shortcuts.

radar
Char vector with three letter radar sindicators. Make sure the radars selected are in accordance to the country selection

directory
Char defining the location to store the downloaded zip folders and unzip into the default folder structure

Examples

my_path <- "~/my/directory/"
# NOT RUN {
download_vpfiles("2016-10-01", "2016-11-30", c("be"),
c("jab", "wid"), directory = my_path)
# }