Gaussian beam profile as a function of height

Normalized altitudinal pattern of radiated energy as a function of altitude at a given distance from the radar. antenna and height should be given in reference to the same reference plane (e.g. ground level or sea level).

Usage

gaussian_beam_profile(
  height,
  range,
  elev,
  antenna = 0,
  beam_angle = 1,
  k = 4/3,
  lat = 35,
  re = 6378,
  rp = 6357
)

Arguments

height

Numeric. Height in m.

range

Numeric. Slant range, i.e. the length of the skywave path between target and the radar antenna, in m.

elev

Numeric. Beam elevation, in degrees.

antenna

Numeric. Height of the centre of the radar antenna, in m.

beam_angle

Numeric. Beam opening angle in degrees, typically the angle between the half-power (-3 dB) points of the main lobe.

k

Numeric. Standard refraction coefficient.

lat

Numeric. Geodetic latitude of the radar, in degrees.

re

Numeric. Earth equatorial radius, in km.

rp

Numeric. Earth polar radius, in km.

Value

numeric.

Details

Beam profile is calculated using beam_height and beam_width. height and antenna should be given in reference to the same reference plane (e.g. ground level or sea level)

See also