Get elevation angles of a scan (scan) or polar volume (pvol)

Gives the elevation angle of a single scan, or the elevation angles of all scans within a polar volume

get_elevation_angles(x)

# S3 method for pvol
get_elevation_angles(x)

# S3 method for scan
get_elevation_angles(x)

# S3 method for param
get_elevation_angles(x)

Arguments

x
A pvol or scan object.

Value

elevation in degrees

Methods (by class)

pvol: Elevation angles of all scans in a polar volume.

scan: Elevation angle of a scan.

param: Elevation angle of a scan parameter.

Examples

# load a polar volume
pvol <- system.file("extdata", "volume.h5", package="bioRad")
vol <- read_pvolfile(pvol)
# elevations for the scans in the volume
get_elevation_angles(vol)

#>  [1]  0.5  0.5  1.0  1.5  2.0  2.5  4.0  8.0 14.0 24.0 40.0

# extract the first scan:
scan <- vol$scans[[1]]
# elevation angle of the scan:
get_elevation_angles(scan)

#> [1] 0.5