Migration traffic

Total migration traffic, which is calculated by time-integration of migration traffic rates. Migration traffic gives the number of individuals that have passed per km perpendicular to the migratory direction at the position of the radar for the full period of the time series within the specified altitude band.

mt(x, alt.min = 0, alt.max = Inf, alpha = NA, interval.max = Inf)

Arguments

x
an object inhereting from class 'vpts'

alt.min
minimum altitude in m

alt.max
maximum altitude in m

alpha
(optional) migratory direction of interest in clockwise degrees from north, otherwise NA

interval.max
maximum time interval belonging to a single profile in seconds. Traffic rates are set to zero at times t for which no profiles can be found within the period t-interval.max/2 to t+interval.max/2.

Value

a numeric value equal to migration traffic in number of individuals / km

Examples

# get example time series object:
data(example_vpts)
example_vpts

#>                    Irregular time series of vertical profiles (class vpts)
#> 
#>            radar:  KBGM 
#>       # profiles:  6135 
#> time range (UTC):  2016-09-01 00:02:00 - 2016-09-30 23:58:00 
#>    time step (s):  min: 180     max:  16320 

# total migration traffic in full altitude band
mt(example_vpts)

#> Warning: 'mt' is deprecated.
#> Use 'integrate_profile' instead.
#> See help("Deprecated")

#> Warning: 'mt' has been moved to the 'mt' column in the output of integrate_profile()

#> Warning: 'cmt' is deprecated.
#> Use 'integrate_profile' instead.
#> See help("Deprecated")

#> Warning: 'cmt' has been moved to the 'mt' column in the output of integrate_profile()

#> [1] 717536.2

# total migration traffic in 0-1000 meter band
mt(example_vpts, alt.min = 0, alt.max = 1000)

#> Warning: 'mt' is deprecated.
#> Use 'integrate_profile' instead.
#> See help("Deprecated")

#> Warning: 'mt' has been moved to the 'mt' column in the output of integrate_profile()

#> Warning: 'cmt' is deprecated.
#> Use 'integrate_profile' instead.
#> See help("Deprecated")

#> Warning: 'cmt' has been moved to the 'mt' column in the output of integrate_profile()

#> [1] 594508.7