Migration traffic rate
Source: R/bioRad.R

Migration traffic rate (MTR) for an altitude layer, defined as the number of targets crossing a 1 km line perpendicular to the migratory movement per hour

mtr(x, alt.min = 0, alt.max = Inf, alpha = NA)

Arguments

x
a vp, vplist or vpts object

alt.min
minimum altitude in m

alt.max
maximum altitude in m

alpha
(optional) migratory direction of interest in clockwise degrees from north, otherwise NA

Value

an atomic vector of migration traffic rates in individuals/km/hour

Details

Migration traffic rate (MTR) for an altitude layer is a flux measure, defined as the number of targets crossing a unit of transect per hour.

The transect direction is set by the angle alpha. When alpha=NA, the transect runs perpendicular to the measured migratory direction. mtr then equals the number of crossing targets per km transect per hour, for a transect kept perpendicular to the measured migratory movement at all times and altitudes. In this case mtr is always a positive quantity, defined as: $$mtr = \sum_i dens_i ff_i \Delta h$$ with the sum running over all altitude layers between alt.min and alt.max, \(dens_i\) the bird density, \(ff_i\) the ground speed at altitude layer i, and \(\Delta h\) the altitude layer width.

If alpha is given a numeric value, the transect is taken perpendicular to the direction alpha, and the number of crossing targets per hour per km transect is calculated as:

$$mtr = \sum_i dens_i ff_i \cos(dd_i-alpha) \Delta h$$ with \(dd_i\) the migratory direction at altitude i.

Note that this equation evaluates to the previous equation when alpha equals \(dd_i\). In this definition mtr is a traditional flux into a direction of interest. Targets moving into the direction alpha contribute positively to mtr, while targets moving in the opposite direction contribute negatively to mtr. Therefore mtr can be both positive or negative, depending on the definition of alpha.

This is a wrapper function for vintegrate, extracting only the migration traffic rate data.

Examples

# NOT RUN {
### MTR for a single vertical profile ###
mtr(VP)
### MTRs for a time series of vertical profiles ###
# load example time series
data(VPTS)
# print migration traffic rates:
mtr(VPTS)
# to plot migration traffic rate data, use vintegrate:
plot(vintegrate(VPTS),quantity="mtr")
# }