Plot a plan position indicator (PPI)
Source: R/bioRad.ppi.R

Plots a plan position indicator (PPI) generated with ppi using ggplot

# S3 method for ppi
plot(x, param, xlim, ylim, zlim = c(-20, 20), ratio = 1, ...)

Arguments

x
an object of class 'ppi'

param
the scan parameter to plot, see details below

xlim
range of x values to plot

ylim
range of y values to plot

zlim
the range of parameter values to plot

ratio
aspect ratio between x and y scale

...
arguments passed to low level ggplot function

Details

Available scan parameters for plotting can by printed to screen by summary(x). Commonly available parameters are:

"DBZH", "DBZ"

(Logged) reflectivity factor [dBZ]

"VRADH", "VRAD"

Radial velocity [m/s]. Radial velocities towards the radar are negative, while radial velocities away from the radar are positive

"RHOHV"

Correlation coefficient [unitless]. Correlation between vertically polarized and horizontally polarized reflectivity factor

"PHIDP"

Differential phase [degrees]

"ZDR"

(Logged) differential reflectivity [dB]

The scan parameters are named according to the OPERA data information model (ODIM), see Table 16 in the ODIM specification.

Examples

# NOT RUN {
# load an example scan:
data(SCAN)
# print to screen the available scan parameters:
summary(SCAN)
# make ppi for the scan
ppi=ppi(SCAN)
# plot the first scan parameter, which in this case is "VRADH":
plot(ppi)
# plot the reflectivity quantity:
plot(ppi,param="DBZH")
# change the range of reflectivities to plot to -30 to 50 dBZ:
plot(ppi,param="DBZH",zlim=c(-30,50))
# }