Make a plan position indicator (ppi)
Source: R/bioRad.ppi.R

Make a plan position indicator (ppi)

ppi(x, cellsize = 500, range.max = 50000, project = F, latlim = NULL,
  lonlim = NULL)

# S3 method for param
ppi(x, cellsize = 500, range.max = 50000, project = F,
  latlim = NULL, lonlim = NULL)

# S3 method for scan
ppi(x, cellsize = 500, range.max = 50000, project = F,
  latlim = NULL, lonlim = NULL)

Arguments

x
an object of class 'param' or 'scan'

cellsize
cartesian grid size in m

range.max
maximum range in m

project
whether to vertically project onto earth's surface

latlim
the range of latitudes to include

lonlim
the range of longitudes to include

...
arguments passed to methods

Value

an object of class 'ppi'.

Details

The returned PPI is in Azimuthal Equidistant Projection.

Methods (by class)

param: ppi for a single scan parameter

scan: multiple ppi's for all scan parameters in a scan

Examples

# NOT RUN {
# load a polar scan example object
data(SCAN)
SCAN
# make PPIs for all scan parameters in the scan:
ppi=ppi(SCAN)
# print summary info for the ppi:
ppi
# copy the first scan parameter of the first scan in the volume to a new object 'param':
param=SCAN$params[[1]]
# make a ppi for the new 'param' object:
ppi=ppi(param)
# print summary info for this ppi:
ppi
# }