Get radar cross section

Gives the currently assumed radar cross section in cm^2.

rcs(x)

# S3 method for vp
rcs(x)

# S3 method for list
rcs(x)

# S3 method for vpts
rcs(x)

# S3 method for vpi
rcs(x)

Arguments

x
A vp, list of vp or vpts object.

Value

a radar cross section in cm^2

Methods (by class)

vp: radar cross section of a vertical profile

list: radar cross sections for a list of vertical profiles

vpts: radar cross section of a time series of vertical profile

vpi: radar cross section of a time series of vertically integrated vertical profile(s)

Examples

# extract RCS for a single vertical profile:
rcs(example_vp)

#> [1] 11