Read a time series of vertical profiles (vpts) from file

Read a time series of vertical profiles (vpts) from file

read_vpts(file, radar, lat, lon, height, wavelength = "C")

Arguments

file
A text file containing the standard output (stdout) generated by vol2bird (or the package function calculate_vp).

radar
A string containing a radar identifier.

lat
numeric. Latitude of the radar in decimal degrees.

lon
numeric. Longitude of the radar in decimal degrees.

height
numeric. Height above sea level of the radar antenna in meters.

wavelength
Radar wavelength in cm, or one of 'C' or 'S' for C-band and S-band radar, respectively, in which case C-band wavelength is assumed to be 5.3 cm and S-band wavelength 10.6 cm

Value

An object inheriting from class vpts, see vpts for details.

Examples

# locate example file:
vptsfile <- system.file("extdata", "example_vpts.txt", package = "bioRad")
# load time series:
ts <- read_vpts(vptsfile, radar = "KBGM", wavelength = "S")
ts

#>                    Irregular time series of vertical profiles (class vpts)
#> 
#>            radar:  KBGM 
#>       # profiles:  11 
#> time range (UTC):  2016-09-02 00:16:00 - 2016-09-02 01:52:00 
#>    time step (s):  min: 540     max:  600