Subset a time series of vertical profiles (vpts)

Select a vertical profile (vp) or a time series of vertical profiles (vpts) by index from a vpts.

# S3 method for vpts
[(x, i)

Arguments

x
A vpts object.

i
Integer. Index/indices specifying which range of vertical profiles to extract.

Value

A vpts object containing a subset of vertical profiles (vp) or a vp object when subsetting a single vertical profile (vp).

Examples

# The example vpts contains 1934 profiles (i.e. datetimes)
dim(example_vpts)

#> [1] 1934   25   15

# Subset vpts to extract 10th profile
example_vpts[10] # A vp object

#>                Vertical profile (class vp)
#> 
#>        radar:  KBGM 
#>       source:  
#> nominal time:  2016-09-01 00:38:00 
#> generated by:   

# Subset vpts to extract the 20th to 100th profile
example_vpts[20:100] # A vpts object with 81 profiles

#>                    Irregular time series of vertical profiles (class vpts)
#> 
#>            radar:  KBGM 
#>       # profiles:  81 
#> time range (UTC):  2016-09-01 01:18:00 - 2016-09-01 07:36:00 
#>    time step (s):  min: 240     max:  420 

# Subset vpts to remove the first 10 profiles
example_vpts[-1:-10] # A vpts object with 10 less profiles

#>                    Irregular time series of vertical profiles (class vpts)
#> 
#>            radar:  KBGM 
#>       # profiles:  1924 
#> time range (UTC):  2016-09-01 00:42:00 - 2016-09-10 11:56:00 
#>    time step (s):  min: 180     max:  16320