Class ppi: a plan position indicator

Class ppi for a plan position indicator, and its associated R base functions.

# S3 method for ppi
summary(object, ...)

is.ppi(x)

# S3 method for ppi
dim(x)

Arguments

object
Object of class ppi.

...
Additional arguments affecting the summary produced.

x
Object of class ppi.

Value

For is.ppi: TRUE if its argument is of class ppi.

For dim.ppi: dimensions of the ppi.

Details

An object of class ppi is a list containing:

data

an object of class SpatialGridDataFrame containing the georeferenced data. Commonly available parameters are:

"DBZH", "DBZ"

(Logged) reflectivity factor [dBZ]

"VRADH", "VRAD"

Radial velocity [m/s]. Radial velocities towards the radar are negative, while radial velocities away from the radar are positive

"RHOHV"

Correlation coefficient [unitless]. Correlation between vertically polarized and horizontally polarized reflectivity factor

"PHIDP"

Differential phase [degrees]

"ZDR"

(Logged) differential reflectivity [dB]

geo

geographic data, a list with:

lat

latitude of the radar [decimal degrees]

lon

longitude of the radar [decimal degrees]

height

height of the radar antenna [metres above sea level]

elangle

radar beam elevation [degrees]

rscale

range bin size [m]

ascale

azimuth bin size [deg]

The geo element of a 'scan' object is a copy of the geo element of its parent scan or scan parameter.