Class 'pvol': polar volume

Class 'pvol': polar volume

# S3 method for pvol
summary(object, ...)

is.pvol(x)

Arguments

object
object of class 'pvol'

...
additional arguments affecting the summary produced.

x
object of class 'pvol'

Value

for is.pvol: TRUE if its argument is of class "pvol"

Details

A polar scan object of class 'pvol' is a list containing:

radar

character string with the radar identifier

datetime

nominal time of the volume [UTC]

scans

a list with scan objects of class 'scan'

attributes

list with the volume's \what, \where and \how attributes

geo

geographic data, a list with:

lat

latitude of the radar [decimal degrees]

lon

longitude of the radar [decimal degrees]

height

height of the radar antenna [metres above sea level]

Examples

# NOT RUN {
# locate example volume file:
pvol <- system.file("extdata", "volume.h5", package="bioRad")
# print the local path of the volume file:
pvol
# load the file:
vol=read.pvol(pvol)
# print summary info for the loaded polar volume:
vol
# print summary info for the scans in the polar volume:
vol$scans
# copy the first scan to a new object 'scan'
scan=vol$scans[[1]]
is.pvol("this is not a polar volume but a string")  #> FALSE
# }