Inspect an integrated profile (vpi)

R base functions for inspecting an integrated profile of biological targets (vpi) object.

Usage

# S3 method for class 'vpi'
summary(object, ...)

is.vpi(x)

Arguments

object

A vpi object.

...

Additional arguments affecting the summary produced.

x

A vpi object.

Value

For summary.vpi(): prints summary of the vpi object.

For is.vpi(): TRUE for an object of class vpi, otherwise FALSE.

Details

A integrated profile of biological targets is a specially classed data.frame generated by the function integrate_profile() with the following quantities:

radar: Radar identifier.

datetime: POSIXct date of each profile in UTC.

vid: Vertically Integrated Density in individuals/km^2. vid is a surface density, whereas dens in vp objects is a volume density.

vir: Vertically Integrated Reflectivity in cm^2/km^2.

mtr: Migration Traffic Rate in individuals/km/h.

rtr: Reflectivity Traffic Rate in cm^2/km/h.

mt: Migration Traffic in individuals/km, cumulated from the start of the time series up to datetime.

rt: Reflectivity Traffic in cm^2/km, cumulated from the start of the time series up to datetime.

ff: Horizontal ground speed in m/s.

dd: Horizontal ground speed direction in degrees.

u: Ground speed component west to east in m/s.

v: Ground speed component south to north in m/s.

height: Mean flight height (height weighted by eta) in m above sea level.

See also

Examples

# Load the example vertical profile time series and integrate to a vpi
vpi <- integrate_profile(example_vpts)

# Check if it is an object of class vpi
is.vpi(vpi)
#> [1] TRUE

# Get summary info
summary(vpi)
#>                Vertically integrated profile(s) (class vpi)
#> 
#>            radar:  KBGM 
#>       # profiles:  1934 
#> time range (UTC):  2016-09-01 00:02:00 - 2016-09-10 11:56:00