Object of class vpts: a time series of vertical profiles

Class for single-site time series of vertical profiles

# S3 method for vpts
summary(object, ...)

is.vpts(x)

# S3 method for vpts
dim(x)

Arguments

object
An object of class vpts.

...
Additional arguments affecting the summary produced.

x
An object of class vpts.

Value

For is.vpts: TRUE if its argument is of class vpts.

For dim.vpts: dimensions of the time series.

Details

An object of class vpts contains time-ordered profiles of a single radar station.

The time series can be regular or irregular, indicated by the regular field

In a regular vpts object the profiles are equally spaced in time. In an irregular vpts object the time steps between profiles are of unequal length.

Irregular time series can be projected onto a regular time grid using the regularize_vpts function.

By contrast, vp objects can be concatenated in a list to combine profiles without time ordering, and profiles of multiple radars.

Data contained in this class object should be accessed with the get_quantity function. Information stored under attributes (see below) can be accessed directly.

An object of class vpts is a list containing

radar

string containing the radar identifier

dates

the N nominal times of the profiles

heights

the M heights of the layers in the profile

daterange

the minimum and maximum nominal time of the profiles in the list

timesteps

time differences between the profiles. Element i gives the time difference between profile i and i+1

data

list of N by M matrices containing the vertical profiles for each quantity. For a description of available quantities, see the data element of the vp class in read_vpfiles

attributes

profile attributes, copied from the first profile contained in x

regular

logical indicating whether the time series is regular or not