Calculate sunrise and sunset

Calculate sunrise and sunset

suntime(lon, lat, date, elev = -0.268, rise = TRUE)

Arguments

lon
longitude in decimal degrees

lat
latitude in decimal degrees

date
date inhereting from class "POSIXt" or a string interpretable by as.Date

elev
sun elevation in degrees

rise
whether to output for rising or setting sun

Value

the moment of sunrise or sunset in UTC time

Details

The angular diameter of the sun is about 0.536 degrees, therefore the moment of sunrise/sunset corresponds to half that elevation at -0.268 degrees.

Note that for a given date and location, sunrise time can be after sunset time, depending on the time difference between the local time and the UTC time zone.

Approximate astronomical formula are used, therefore the moment of sunrise / sunset may be off by a few minutes

Examples

# NOT RUN {
# sunrise in the Netherlands
suntime(5,53,"2016-01-01")
# sunset in the Netherlands
suntime(5,53,"2016-01-01",rise=FALSE)
# civil twilight in Ithaca, NY, today
suntime(-76.5,42.4,Sys.time(),elev=-6)
# }