Deprecated functions in package bioRad

The functions listed below are deprecated or renamed and will be defunct in the near future. When possible, alternative functions with similar functionality are mentioned.

beamheight(range, elev, k = 4/3, lat = 35, re = 6378, rp = 6357)

beamwidth(range, angle = 1)

bind(x, ...)

vol2bird(...)

night(x, ..., elev = -0.268)

day(x, elev = -0.268)

composite(...)

dbz2eta(...)

eta2dbz(...)

checkDocker(...)

updateDocker(...)

basemap(...)

download_vp(...)

elangle(...)

fetch(...)

getscan(...)

vintegrate(...)

rsl2odim(...)

ppi(...)

h5ODIMobject(...)

read.pvol(...)

readvp(...)

readvp.list(...)

readvp.table(...)

regularize(...)

sd_vvp(...)

sd_vvp(x) <- value

retrieve_vp_paths(...)

suntime(..., rise = TRUE)

is.vplist(...)

vpts(...)

VP(...)

VPTS(...)

SCAN(...)

<code>beamheight</code>

Use beam_height instead.

<code>beamwidth</code>

Use beam_width instead.

<code>bind</code>

Use bind_into_vpts instead.

<code>vol2bird</code>

Use calculate_vp instead.

<code>night</code>

Use check_night instead.

<code>day</code>

Use check_night instead.

<code>composite</code>

Use composite_ppi instead.

<code>dbz2eta</code>

Use dbz_to_eta instead.

<code>eta2dbz</code>

Use eta_to_dbz instead.

<code>checkDocker</code>

Use check_docker instead.

<code>updateDocker</code>

Use update_docker instead.

<code>basemap</code>

Use download_basemap instead.

<code>download_vp</code>

Use download_vpfiles instead.

<code>elangle</code>

Use get_angles instead.

<code>fetch</code>

Use get_quantity instead.

<code>getscan</code>

Use get_scan instead.

<code>vintegrate</code>

Use integrate_profile instead.

<code>rsl2odim</code>

Use nexrad_to_odim instead.

<code>ppi</code>

Use project_as_ppi instead.

<code>h5ODIMobject</code>

Use get_odim_object_type instead.

<code>read.pvol</code>

Use read_pvolfile instead.

<code>readvp</code>

Use read_vpfiles instead.

<code>readvp.list</code>

Use read_vpfiles instead.

<code>readvp.table</code>

Use read_vpts instead.

<code>regularize</code>

Use regularize_vpts instead.

<code>sd_vvp</code>

Use rvsd instead.

<code>sd_vvp&lt;-</code>

Use rvsd<- instead.

<code>retrieve_vp_paths</code>

Use select_vpfiles instead.

<code>suntime</code>

Use sunrise or sunset instead.

<code>is.vplist</code>

The vplist object is deprecated. Use a regular list of vp objects (c(vp, vp)) instead.

<code>vpts</code>

Use bind_to_vpts instead.

<code>VP</code>

Use example_vp instead.

<code>VPTS</code>

Use example_vpts instead.

<code>SCAN</code>

Use example_scan instead.