Deprecated functions and data in package bioRad

The functions and data listed below are deprecated or renamed and will be defunct in the near future. When possible, alternative functions with similar functionality are mentioned.

Total migration traffic, which is calculated by time-integration of migration traffic rates. Migration traffic gives the number of individuals that have passed per km perpendicular to the migratory direction at the position of the radar for the full period of the time series within the specified altitude band.

Migration traffic rate (MTR) for an altitude layer, defined as the number of targets crossing a 1 km line perpendicular to the migratory movement per hour.

Cumulative migration traffic is calculated as the cumulative sum of the migration traffic within each time step of a time series. Cumulative migration traffic gives the number of individuals that have passed per km perpendicular to the migratory direction at the position of the radar as a function of time from the start of time series within the specified altitude band.

basemap(...)

beamheight(range, elev, k = 4/3, lat = 35, re = 6378, rp = 6357)

beamwidth(range, angle = 1)

bind(x, ...)

checkDocker(...)

composite(x, param = "DBZH", cells.dim = c(100, 100))

day(x, ..., elev = -0.268)

dbz2eta(dbz, wavelength, Km = 0.93)

download_vp(start_date, end_date, country, radar, localpath = ".")

elangle(...)

eta2dbz(eta, wavelength, Km = 0.93)

fetch(...)

getscan(x, angle)

h5ODIMobject(filename)

is.vplist(...)

mt(x, alt.min = 0, alt.max = Inf, alpha = NA, interval.max = Inf)

mtr(x, alt.min = 0, alt.max = Inf, alpha = NA)

cmt(x, alt.min = 0, alt.max = Inf, alpha = NA, interval.max = Inf)

night(x, ..., elev = -0.268)

ppi(x, cellsize = 500, range.max = 50000, project = FALSE,
  latlim = NULL, lonlim = NULL)

read.pvol(filename, param = c("DBZH", "VRADH", "VRAD", "RHOHV", "ZDR",
  "PHIDP", "CELL"), sort = TRUE, lat, lon, height, elangle.min = 0,
  elangle.max = 90, verbose = TRUE, mount = dirname(filename))

readvp(...)

readvp.list(...)

readvp.table(...)

regularize(ts, interval = "auto", t.min = ts$daterange[1],
  t.max = ts$daterange[2], units = "mins", fill = FALSE,
  verbose = TRUE)

retrieve_vp_paths(path, start_date, end_date, country = NULL,
  radar = NULL)

rsl2odim(vol.in, vol.out, verbose = FALSE, mount = dirname(vol.in))

sd_vvp(x)

suntime(..., rise = TRUE)

updateDocker(...)

vintegrate(...)

vol2bird(vol.in, vp.out = "", vol.out = "", autoconf = FALSE,
  verbose = FALSE, mount = dirname(vol.in), sd_vvp_threshold = 2,
  rcs = 11, dualpol = FALSE, rhohv = 0.95, elev.min = 0,
  elev.max = 90, azim.min = 0, azim.max = 360, range.min = 5000,
  range.max = 25000, nlayer = 20L, hlayer = 200, dealias = TRUE,
  nyquist.min = if (dealias) 5 else 25, dbz_quantity = "DBZH")

vpts(...)

Arguments

x
A vpts.

alt.min
Minimum altitude in m.

alt.max
Maximum altitude in m.

alpha
(optional) migratory direction of interest in clockwise degrees from north, otherwise NA.

interval.max
Maximum time interval belonging to a single profile in seconds. Traffic rates are set to zero at times t for which no profiles can be found within the period t-interval.max/2 to t+interval.max/2.

x
A vp, list of vp objects or vpts object.

x
A vpts.

interval.max
Maximum time interval belonging to a single profile in seconds. Traffic rates are set to zero at times t for which no profiles can be found within the period t-interval.max/2 to t+interval.max/2.

Value

A numeric value equal to migration traffic in number of individuals / km.

An atomic vector of migration traffic rates in individuals/km/hour.

Atomic vector with (cumulative) migration traffic in number of individuals / km.

Details

Migration traffic rate (MTR) for an altitude layer is a flux measure, defined as the number of targets crossing a unit of transect per hour.

The transect direction is set by the angle alpha. When alpha=NA, the transect runs perpendicular to the measured migratory direction. mtr then equals the number of crossing targets per km transect per hour, for a transect kept perpendicular to the measured migratory movement at all times and altitudes. In this case mtr is always a positive quantity, defined as:

$$mtr = \sum_i dens_i ff_i \Delta h$$

with the sum running over all altitude layers between alt.min and alt.max, \(dens_i\) the bird density, \(ff_i\) the ground speed at altitude layer i, and \(\Delta h\) the altitude layer width.

If alpha is given a numeric value, the transect is taken perpendicular to the direction alpha, and the number of crossing targets per hour per km transect is calculated as:

$$mtr = \sum_i dens_i ff_i \cos(dd_i-alpha) \Delta h$$ with \(dd_i\) the migratory direction at altitude i.

Note that this equation evaluates to the previous equation when alpha equals \(dd_i\).

In this definition mtr is a traditional flux into a direction of interest. Targets moving into the direction alpha contribute positively to mtr, while targets moving in the opposite direction contribute negatively to mtr. Therefore mtr can be both positive or negative, depending on the definition of alpha.

This is a wrapper function for integrate_profile, extracting only the migration traffic rate data.

basemap

Use download_basemap instead.

beamheight

Use beam_height instead.

beamwidth

Use beam_width instead.

bind

Use bind_into_vpts instead.

checkDocker

Use check_docker instead.

composite

Use composite_ppi instead.

day

Use check_night instead.

dbz2eta

Use dbz_to_eta instead.

download_vp

Use download_vpfiles instead.

elangle

Use get_elevation_angles instead.

eta2dbz

Use eta_to_dbz instead.

fetch

Use get_quantity instead.

getscan

Use get_scan instead.

h5ODIMobject

Use get_odim_object_type instead.

is.vplist

The vplist object is deprecated. Use a regular list of vp objects (c(vp, vp)) instead.

mt

mt is deprecated as a separate function. Migration traffic is now included in the output of integrate_profile as column mt.

mtr

mtr is deprecated as a separate function. Migration traffic rate is now included in the output of integrate_profile as column mtr.

cmt

cmt is deprecated as a separate function. Cumulative migration traffic is now included in the output of integrate_profile as column mt, which can be summed to get the cumulative migration traffic.

night

Use check_night instead.

ppi

Use project_as_ppi instead.

read.pvol

Use read_pvolfile instead.

readvp

Use read_vpfiles instead.

readvp.list

Use read_vpfiles instead.

readvp.table

Use read_vpts instead.

regularize

Use regularize_vpts instead.

retrieve_vp_paths

Use select_vpfiles instead.

rsl2odim

Use nexrad_to_odim instead.

sd_vvp

Use sd_vvp_threshold instead.

suntime

Use sunrise or sunset instead.

updateDocker

Use update_docker instead.

vintegrate

Use integrate_profile instead.

vol2bird

Use calculate_vp instead.

vpts

Use bind_into_vpts instead.

VP

Use example_vp instead.

VPTS

Use example_vpts instead.

SCAN

Use example_scan instead.

See also

Examples

# get example time series of vertical profiles:
data(example_vpts)
example_vpts

#>                    Irregular time series of vertical profiles (class vpts)
#> 
#>            radar:  KBGM 
#>       # profiles:  1934 
#> time range (UTC):  2016-09-01 00:02:00 - 2016-09-10 11:56:00 
#>    time step (s):  min: 180     max:  16320 

# total migration traffic in full altitude band:

# NOT RUN {
mt(example_vpts)
# }
# total migration traffic in 0-1000 meter band:

# NOT RUN {
mt(example_vpts, alt.min = 0, alt.max = 1000)
# }
# MTR for a single vertical profile:

# NOT RUN {
mtr(example_vp)
# }
# MTRs for a time series of vertical profiles:
data(example_vpts)
# print migration traffic rates:
# NOT RUN {
mtr(example_vpts)
# }
# to plot migration traffic rate data, use integrate_profile:

# NOT RUN {
plot(integrate_profile(example_vpts), quantity = "mtr")
# }
# get example time series of vertical profiles:
data(example_vpts)

# print cumulative migration traffic to console:
# NOT RUN {
cmt(example_vpts)
# }

# plot cumulative migration traffic:

# NOT RUN {
plot(cmt(example_vpts), type = "l", xlab = "time" ,
ylab = "CMT [birds/km]")
# }