Calculate a vertical profile (vp) from a polar volume (pvol)

Calculates a vertical profile of biological scatterers (vp) from a polar volume (pvol) using the algorithm vol2bird (Dokter et al. 2011).

calculate_vp(
  file,
  vpfile = "",
  pvolfile_out = "",
  autoconf = FALSE,
  verbose = FALSE,
  mount = dirname(file[1]),
  sd_vvp_threshold,
  rcs = 11,
  dual_pol = FALSE,
  rho_hv = 0.95,
  elev_min = 0,
  elev_max = 90,
  azim_min = 0,
  azim_max = 360,
  range_min = 5000,
  range_max = 35000,
  n_layer = 20L,
  h_layer = 200,
  dealias = TRUE,
  nyquist_min = if (dealias) 5 else 25,
  dbz_quantity = "DBZH",
  mistnet = FALSE,
  mistnet_elevations = c(0.5, 1.5, 2.5, 3.5, 4.5),
  local_install,
  local_mistnet,
  pvolfile
)

Arguments

file
string or a vector of strings with radar file(s) for a radar polar volume. Provide either a single file containing a polar volume, or multiple files with single scans/sweeps. Data format should be either ODIM format, which is the implementation of the OPERA data information model in HDF5 format, or a format supported by the RSL library, or Vaisala IRIS (IRIS RAW) format.

vpfile
character. Filename for the vertical profile to be generated in ODIM HDF5 format (optional).

pvolfile_out
character. Filename for the polar volume to be generated in ODIM HDF5 format (optional, e.g. for converting RSL formats to ODIM).

autoconf
logical. When TRUE, default optimal configuration settings are selected automatically, and other user settings are ignored.

verbose
logical. When TRUE, pipe Docker stdout to R console. On Windows always TRUE.

mount
character. String with the mount point (a directory path) for the Docker container.

sd_vvp_threshold
numeric. Lower threshold in radial velocity standard deviation (profile quantity sd_vvp) in m/s. Biological signals with sd_vvp < sd_vvp_threshold are set to zero. Defaults to 2 m/s for C-band radars and 1 m/s for S-band radars if not specified.

rcs
numeric. Radar cross section per bird in cm^2.

dual_pol
logical. When TRUE use dual-pol mode, in which meteorological echoes are filtered using the correlation coefficient rho_hv. When FALSE use single polarization mode based only on reflectivity and radial velocity quantities.

rho_hv
numeric. Lower threshold in correlation coefficient used to filter meteorological scattering.

elev_min
numeric. Minimum scan elevation in degrees.

elev_max
numeric. Maximum scan elevation in degrees.

azim_min
numeric. Minimum azimuth in degrees clockwise from north.

azim_max
numeric. Maximum azimuth in degrees clockwise from north.

range_min
numeric. Minimum range in m.

range_max
numeric. Maximum range in m.

n_layer
numeric. Number of altitude layers in the profile.

h_layer
numeric. Width of altitude layers in meter.

dealias
logical. Whether to dealias radial velocities; this should typically be done when the scans in the polar volume have low Nyquist velocities (below 25 m/s).

nyquist_min
numeric. Minimum Nyquist velocity of scans in m/s for scans to be included in the analysis.

dbz_quantity
character. One of the available reflectivity factor quantities in the ODIM radar data format, e.g. DBZH, DBZV, TH, TV.

mistnet
logical. Whether to use MistNet segmentation model.

local_install
(optional) String with path to local vol2bird installation (e.g. "/your/vol2bird_install_directory/vol2bird/bin/vol2bird"). To use local installation instead of Docker container, see details.

local_mistnet
(optional) String with path to local mistnet segmentation model in PyTorch format (e.g. "/your/path/mistnet_nexrad.pt"), to use local installation instead of Docker container.

pvolfile
deprecated argument renamed to file.

Value

A vertical profile object of class vp. When defined, output files vpfile and pvolfile_out are saved to disk.

Details

Requires a running Docker daemon (unless a local installation of vol2bird is specified with local_install).

Common arguments set by users are file, vpfile, autoconf and mount.

Turn on autoconf to automatically select the optimal parameters for a given radar file. The default for C-band data is to apply rain-filtering in single polarization mode, as well as dual polarization mode when available.

The default for S-band data is to apply precipitation filtering in dual-polarization mode.

Arguments that sometimes require non-default values are: rcs, sd_vvp_threshold, range_max, dual_pol, dealias.

Other arguments are typically left at their defaults.

azim_min and azim_max only affects reflectivity-derived estimates in the profile (DBZH,eta,dens), not radial-velocity derived estimates (u, v, w, ff, dd, sd_vvp), which are estimated on all azimuths at all times. azim_min, azim_max may be set to exclude an angular sector with high ground clutter.

range_max may be extended up to 40,000 m for volumes with low elevations only, in order to extend coverage to higher altitudes.

For altitude layers with a VVP-retrieved radial velocity standard deviation value below the threshold sd_vvp_threshold, the bird density dens is set to zero (see vertical profile vp class). This threshold might be dependent on radar processing settings. Results from validation campaigns so far indicate that 2 m/s is the best choice for this parameter for most C-band weather radars, which is used as the C-band default. For S-band, the default threshold is 1 m/s. The algorithm has been tested and developed for altitude layers with h_layer = 200 m. Smaller widths are not recommended as they may cause instabilities of the volume velocity profiling (VVP) and dealiasing routines, and effectively lead to pseudo-replicated altitude data, since altitudinal patterns smaller than the beam width cannot be resolved.

The default radar cross section (11 cm^2) corresponds to the average value found by Dokter et al. in a calibration campaign of a full migration autumn season in western Europe at C-band. It's value may depend on radar wavelength. rcs will scale approximately \(M^{2/3}\) with M the bird's mass.

Using default values of range_min and range_max is recommended. Ranges closer than 5 km tend to be contaminated by ground clutter, while range gates beyond 35 km become too wide to resolve the default altitude layer width of 200 meter (see beam_width).

For dealiasing, the torus mapping method by Haase et al. is used.

At S-band (radar wavelength ~ 10 cm), currently only dual_pol=TRUE mode is recommended.

On repeated calls of calculate_vp, the Docker container mount can be recycled from one call to the next if subsequent calls share the same mount argument. Re-mounting a Docker container takes time, therefore it is advised to choose a mountpoint that is a parent directory of all volume files to be processed, such that calculate_vp calls are as fast as possible.

If you have installed the vol2bird algorithm locally (not possible on Windows) you can call vol2bird through this local installation (bypassing the Docker container), which will be faster. Simply point local_install to the path of your local vol2bird executable, e.g. "/your/vol2bird_install_directory/vol2bird/bin/vol2bird". Your local vol2bird executable will be called through a bash login shell. LD_LIBRARY_PATH (Linux) or DYLD_LIBRARY_PATH (Mac) should be correctly specified in your .bashrc or .bash_profile file and contain all the required shared libraries by vol2bird. See vol2bird installation pages on Github for details.

When using MistNet with a local vol2bird installation, also point parameter local_mistnet to your local download of the MistNet segmentation model in PyTorch format, e.g. "/your/path/mistnet_nexrad.pt"). The MistNet model can be downloaded at https://s3.amazonaws.com/mistnet/mistnet_nexrad.pt.

References

Dokter et al. (2011) is the main reference for the profiling algorithm (vol2bird) underlying this function. When using the mistnet option, please also cite Lin et al. 2019. When de-aliasing data, please also cite Haase et al. 2004.

Adriaan M. Dokter, Felix Liechti, Herbert Stark, Laurent Delobbe, Pierre Tabary, Iwan Holleman, 2011. Bird migration flight altitudes studied by a network of operational weather radars, Journal of the Royal Society Interface 8 (54), pp. 30--43. https://doi.org/10.1098/rsif.2010.0116

Haase, G. and Landelius, T., 2004. Dealiasing of Doppler radar velocities using a torus mapping. Journal of Atmospheric and Oceanic Technology, 21(10), pp.1566--1573. https://doi.org/10.1175/1520-0426(2004)021<1566:DODRVU>2.0.CO;2

Tsung-Yu Lin, Kevin Winner, Garrett Bernstein, Abhay Mittal, Adriaan M. Dokter Kyle G. Horton, Cecilia Nilsson, Benjamin M. Van Doren, Andrew Farnsworth Frank A. La Sorte, Subhransu Maji, Daniel Sheldon, 2019. MistNet: Measuring historical bird migration in the US using archived weather radar data and convolutional neural networks Methods in Ecology and Evolution 10 (11), pp. 1908--22. https://doi.org/10.1111/2041-210X.13280

Examples

# locate example polar volume file:
pvolfile <- system.file("extdata", "volume.h5", package = "bioRad")

# copy to a home directory with read/write permissions:
file.copy(pvolfile, "~/volume.h5")

#> [1] TRUE

# calculate the profile:
# \dontrun{
profile <- calculate_vp("~/volume.h5")
# print some summary info:
profile

#>                Vertical profile (class vp)
#> 
#>        radar:  seang 
#>       source:  WMO:02606,RAD:SE50,PLC:Angelholm,NOD:seang,ORG:82,CTY:643,CMT:Swedish radar 
#> nominal time:  2015-10-18 18:00:00 
#> generated by:  vol2bird 0.5.0.9167 

# convert profile to a data.frame:
as.data.frame(profile)

#>    radar            datetime       ff        dbz      dens         u          v
#> 1  seang 2015-10-18 18:00:00       NA         NA        NA        NA         NA
#> 2  seang 2015-10-18 18:00:00      NaN  2.8776991  61.27740       NaN        NaN
#> 3  seang 2015-10-18 18:00:00 14.88610  5.5255389 112.74165 -6.467583 -13.407694
#> 4  seang 2015-10-18 18:00:00 14.64004  4.6949205  93.11562 -6.711918 -13.010798
#> 5  seang 2015-10-18 18:00:00 13.30489  0.4530046  35.06153 -6.556170 -11.577423
#> 6  seang 2015-10-18 18:00:00 10.76581 -1.9438103  20.19061 -5.069613  -9.497456
#> 7  seang 2015-10-18 18:00:00 10.33567 -1.8632607  20.56858 -5.671824  -8.640401
#> 8  seang 2015-10-18 18:00:00 12.07231 -2.2290711  18.90703 -7.580208  -9.395803
#> 9  seang 2015-10-18 18:00:00 10.28255 -2.5026581  17.75271 -3.064733  -9.815207
#> 10 seang 2015-10-18 18:00:00       NA         NA        NA        NA         NA
#> 11 seang 2015-10-18 18:00:00       NA         NA        NA        NA         NA
#> 12 seang 2015-10-18 18:00:00       NA         NA        NA        NA         NA
#> 13 seang 2015-10-18 18:00:00       NA         NA        NA        NA         NA
#> 14 seang 2015-10-18 18:00:00       NA         NA        NA        NA         NA
#> 15 seang 2015-10-18 18:00:00       NA         NA        NA        NA         NA
#> 16 seang 2015-10-18 18:00:00       NA         NA        NA        NA         NA
#> 17 seang 2015-10-18 18:00:00       NA         NA        NA        NA         NA
#> 18 seang 2015-10-18 18:00:00       NA         NA        NA        NA         NA
#> 19 seang 2015-10-18 18:00:00       NA         NA        NA        NA         NA
#> 20 seang 2015-10-18 18:00:00       NA         NA        NA        NA         NA
#>    gap          w n_dbz       dd    n      DBZH height n_dbz_all       eta
#> 1    1         NA     0       NA    0        NA      0         0        NA
#> 2    1        NaN   817      NaN  536  4.873434    200     10602  674.0515
#> 3    0   6.552267  4231 205.7517 2116  8.229273    400     17421 1240.1581
#> 4    0 -17.817648  4085 207.2880 1866  4.437433    600      8098 1024.2719
#> 5    0 -12.915339  3897 209.5224 1155  1.342218    800      7801  385.6769
#> 6    0  30.602781  4369 208.0927  744 -0.492405   1000      7829  222.0967
#> 7    0  24.725313  2276 213.2821  450 -1.101062   1200      3206  226.2544
#> 8    0  57.809780  1694 218.8954  198 -1.102251   1400      2851  207.9773
#> 9    0  36.869717  2045 197.3406  158 10.828478   1600      3213  195.2798
#> 10   1         NA     0       NA    0        NA   1800         0        NA
#> 11   1         NA     0       NA    0        NA   2000         0        NA
#> 12   1         NA     0       NA    0        NA   2200         0        NA
#> 13   1         NA     0       NA    0        NA   2400         0        NA
#> 14   1         NA     0       NA    0        NA   2600         0        NA
#> 15   1         NA     0       NA    0        NA   2800         0        NA
#> 16   1         NA     0       NA    0        NA   3000         0        NA
#> 17   1         NA     0       NA    0        NA   3200         0        NA
#> 18   1         NA     0       NA    0        NA   3400         0        NA
#> 19   1         NA     0       NA    0        NA   3600         0        NA
#> 20   1         NA     0       NA    0        NA   3800         0        NA
#>      sd_vvp n_all     lat     lon height_antenna   day             sunrise
#> 1        NA     0 56.3675 12.8517            209 FALSE 2015-10-18 05:50:07
#> 2  3.271695  4351 56.3675 12.8517            209 FALSE 2015-10-18 05:50:07
#> 3  3.220699  5304 56.3675 12.8517            209 FALSE 2015-10-18 05:50:07
#> 4  3.316308  3144 56.3675 12.8517            209 FALSE 2015-10-18 05:50:07
#> 5  3.897997  2043 56.3675 12.8517            209 FALSE 2015-10-18 05:50:07
#> 6  4.067675  1380 56.3675 12.8517            209 FALSE 2015-10-18 05:50:07
#> 7  4.837257   747 56.3675 12.8517            209 FALSE 2015-10-18 05:50:07
#> 8  4.795825   483 56.3675 12.8517            209 FALSE 2015-10-18 05:50:07
#> 9  4.506003   315 56.3675 12.8517            209 FALSE 2015-10-18 05:50:07
#> 10       NA     0 56.3675 12.8517            209 FALSE 2015-10-18 05:50:07
#> 11       NA     0 56.3675 12.8517            209 FALSE 2015-10-18 05:50:07
#> 12       NA     0 56.3675 12.8517            209 FALSE 2015-10-18 05:50:07
#> 13       NA     0 56.3675 12.8517            209 FALSE 2015-10-18 05:50:07
#> 14       NA     0 56.3675 12.8517            209 FALSE 2015-10-18 05:50:07
#> 15       NA     0 56.3675 12.8517            209 FALSE 2015-10-18 05:50:07
#> 16       NA     0 56.3675 12.8517            209 FALSE 2015-10-18 05:50:07
#> 17       NA     0 56.3675 12.8517            209 FALSE 2015-10-18 05:50:07
#> 18       NA     0 56.3675 12.8517            209 FALSE 2015-10-18 05:50:07
#> 19       NA     0 56.3675 12.8517            209 FALSE 2015-10-18 05:50:07
#> 20       NA     0 56.3675 12.8517            209 FALSE 2015-10-18 05:50:07
#>                 sunset
#> 1  2015-10-18 15:56:28
#> 2  2015-10-18 15:56:28
#> 3  2015-10-18 15:56:28
#> 4  2015-10-18 15:56:28
#> 5  2015-10-18 15:56:28
#> 6  2015-10-18 15:56:28
#> 7  2015-10-18 15:56:28
#> 8  2015-10-18 15:56:28
#> 9  2015-10-18 15:56:28
#> 10 2015-10-18 15:56:28
#> 11 2015-10-18 15:56:28
#> 12 2015-10-18 15:56:28
#> 13 2015-10-18 15:56:28
#> 14 2015-10-18 15:56:28
#> 15 2015-10-18 15:56:28
#> 16 2015-10-18 15:56:28
#> 17 2015-10-18 15:56:28
#> 18 2015-10-18 15:56:28
#> 19 2015-10-18 15:56:28
#> 20 2015-10-18 15:56:28

# }

# clean up:
file.remove("~/volume.h5")

#> [1] TRUE