Look up day of year (doy) or night of year (noy)

Returns the day of year (doy) or night of year (noy) number for datetimes and various bioRad objects. The first night of the year is the night with datetime Jan 01 00:00:00 in the local time zone, so sunset on Jan 1 occurs on the second night of the year and New Years Eve on Dec 31 occurs on the first night of the new year.

doy(x, ..., method = "fast")

noy(x, ..., method = "fast")

# S3 method for default
doy(x, lon, lat, ..., method = "fast")

# S3 method for default
noy(x, lon, lat, ..., method = "fast")

# S3 method for vp
doy(x, ..., method = "fast")

# S3 method for vp
noy(x, ..., method = "fast")

# S3 method for vpts
doy(x, ..., method = "fast")

# S3 method for vpts
noy(x, ..., method = "fast")

# S3 method for vpi
doy(x, ..., method = "fast")

# S3 method for vpi
noy(x, ..., method = "fast")

# S3 method for pvol
doy(x, ..., method = "fast")

# S3 method for pvol
noy(x, ..., method = "fast")

Arguments

x

A pvol, vp, vpts, or vpi object, or a base::as.POSIXct datetime.

...

Optional lat, lon arguments.

method

Method by which to do the time zone lookup. Either fast (default) or accurate. See lutz::tz_lookup_coords].

lon

Numeric. Longitude in decimal degrees.

lat

Numeric. Latitude in decimal degrees.

Examples

# Get day of year of a vp object
noy(example_vp)
#> [1] 292

# Get night of year of a vp object
noy(example_vp)
#> [1] 292

# Get night of year of a vpts object
noy(example_vpts)
#>    [1] 245 245 245 245 245 245 245 245 245 245 245 245 245 245 245 245 245 245
#>   [19] 245 245 245 245 245 245 245 245 245 245 245 245 245 245 245 245 245 245
#>   [37] 245 245 245 245 245 245 245 245 245 245 245 245 245 245 245 245 245 245
#>   [55] 245 245 245 245 245 245 245 245 245 245 245 245 245 245 245 245 245 245
#>   [73] 245 245 245 245 245 245 245 245 245 245 245 245 245 245 245 245 245 245
#>   [91] 245 245 245 245 245 245 245 245 245 245 245 245 245 245 245 245 245 245
#>  [109] 245 245 245 245 245 245 245 245 245 245 245 245 245 245 245 245 245 245
#>  [127] 245 245 245 245 245 245 245 245 245 245 245 245 245 245 245 245 245 245
#>  [145] 245 245 245 245 245 245 245 245 245 245 245 245 245 245 245 245 245 245
#>  [163] 245 245 245 245 245 245 245 245 245 245 245 245 245 245 245 245 245 245
#>  [181] 245 245 245 245 245 245 245 245 245 245 245 245 245 245 245 245 245 245
#>  [199] 245 245 245 245 245 245 245 245 245 245 245 245 245 245 245 245 245 245
#>  [217] 245 245 245 245 245 245 245 245 246 246 246 246 246 246 246 246 246 246
#>  [235] 246 246 246 246 246 246 246 246 246 246 246 246 246 246 246 246 246 246
#>  [253] 246 246 246 246 246 246 246 246 246 246 246 246 246 246 246 246 246 246
#>  [271] 246 246 246 246 246 246 246 246 246 246 246 246 246 246 246 246 246 246
#>  [289] 246 246 246 246 246 246 246 246 246 246 246 246 246 246 246 246 246 246
#>  [307] 246 246 246 246 246 246 246 246 246 246 246 246 246 246 246 246 246 246
#>  [325] 246 246 246 246 246 246 246 246 246 246 246 246 246 246 246 246 246 246
#>  [343] 246 246 246 246 246 246 246 246 246 246 246 246 246 246 246 246 246 246
#>  [361] 246 246 246 246 246 246 246 246 246 246 246 246 246 246 246 246 246 246
#>  [379] 246 246 246 246 247 247 247 247 247 247 247 247 247 247 247 247 247 247
#>  [397] 247 247 247 247 247 247 247 247 247 247 247 247 247 247 247 247 247 247
#>  [415] 247 247 247 247 247 247 247 247 247 247 247 247 247 247 247 247 247 247
#>  [433] 247 247 247 247 247 247 247 247 247 247 247 247 247 247 247 247 247 247
#>  [451] 247 247 247 247 247 247 247 247 247 247 247 247 247 247 247 247 247 247
#>  [469] 247 247 247 247 247 247 247 247 247 247 247 247 247 247 247 247 247 247
#>  [487] 247 247 247 247 247 247 247 247 247 247 247 247 247 247 247 247 247 247
#>  [505] 247 247 247 247 247 247 247 247 247 247 247 247 247 247 247 247 247 247
#>  [523] 247 247 247 247 247 247 247 247 247 248 248 248 248 248 248 248 248 248
#>  [541] 248 248 248 248 248 248 248 248 248 248 248 248 248 248 248 248 248 248
#>  [559] 248 248 248 248 248 248 248 248 248 248 248 248 248 248 248 248 248 248
#>  [577] 248 248 248 248 248 248 248 248 248 248 248 248 248 248 248 248 248 248
#>  [595] 248 248 248 248 248 248 248 248 248 248 248 248 248 248 248 248 248 248
#>  [613] 248 248 248 248 248 248 248 248 248 248 248 248 248 248 248 248 248 248
#>  [631] 248 248 248 248 248 248 248 248 248 248 248 248 248 248 248 248 248 248
#>  [649] 248 248 248 248 248 248 248 248 248 248 248 248 248 248 248 248 248 248
#>  [667] 248 248 248 248 248 248 248 248 248 248 248 248 248 249 249 249 249 249
#>  [685] 249 249 249 249 249 249 249 249 249 249 249 249 249 249 249 249 249 249
#>  [703] 249 249 249 249 249 249 249 249 249 249 249 249 249 249 249 249 249 249
#>  [721] 249 249 249 249 249 249 249 249 249 249 249 249 249 249 249 249 249 249
#>  [739] 249 249 249 249 249 249 249 249 249 249 249 249 249 249 249 249 249 249
#>  [757] 249 249 249 249 249 249 249 249 249 249 249 249 249 249 249 249 249 249
#>  [775] 249 249 249 249 249 249 249 249 249 249 249 249 249 249 249 249 249 249
#>  [793] 249 249 249 249 249 249 249 249 249 249 249 249 249 249 249 249 249 249
#>  [811] 249 249 249 249 249 249 249 250 250 250 250 250 250 250 250 250 250 250
#>  [829] 250 250 250 250 250 250 250 250 250 250 250 250 250 250 250 250 250 250
#>  [847] 250 250 250 250 250 250 250 250 250 250 250 250 250 250 250 250 250 250
#>  [865] 250 250 250 250 250 250 250 250 250 250 250 250 250 250 250 250 250 250
#>  [883] 250 250 250 250 250 250 250 250 250 250 250 250 250 250 250 250 250 250
#>  [901] 250 250 250 250 250 250 250 250 250 250 250 250 250 250 250 250 250 250
#>  [919] 250 250 250 250 250 250 250 250 250 250 250 250 250 250 250 250 250 250
#>  [937] 250 250 250 250 250 250 250 250 250 250 250 250 250 251 251 251 251 251
#>  [955] 251 251 251 251 251 251 251 251 251 251 251 251 251 251 251 251 251 251
#>  [973] 251 251 251 251 251 251 251 251 251 251 251 251 251 251 251 251 251 251
#>  [991] 251 251 251 251 251 251 251 251 251 251 251 251 251 251 251 251 251 251
#> [1009] 251 251 251 251 251 251 251 251 251 251 251 251 251 251 251 251 251 251
#> [1027] 251 251 251 251 251 251 251 251 251 251 251 251 251 251 251 251 251 251
#> [1045] 251 251 251 251 251 251 251 251 251 251 251 251 251 251 251 251 251 251
#> [1063] 251 251 251 251 251 251 251 251 251 251 251 251 251 251 251 251 251 251
#> [1081] 251 251 251 251 251 251 251 251 251 251 251 251 251 251 251 251 251 251
#> [1099] 251 251 251 251 251 251 251 251 251 251 251 251 251 251 251 251 251 251
#> [1117] 251 251 251 251 251 251 251 251 251 252 252 252 252 252 252 252 252 252
#> [1135] 252 252 252 252 252 252 252 252 252 252 252 252 252 252 252 252 252 252
#> [1153] 252 252 252 252 252 252 252 252 252 252 252 252 252 252 252 252 252 252
#> [1171] 252 252 252 252 252 252 252 252 252 252 252 252 252 252 252 252 252 252
#> [1189] 252 252 252 252 252 252 252 252 252 252 252 252 252 252 252 252 252 252
#> [1207] 252 252 252 252 252 252 252 252 252 252 252 252 252 252 252 252 252 252
#> [1225] 252 252 252 252 252 252 252 252 252 252 252 252 252 252 252 252 252 252
#> [1243] 252 252 252 252 252 252 252 252 252 252 252 252 252 252 252 252 252 252
#> [1261] 252 252 252 252 252 252 252 252 252 252 252 252 252 252 252 252 252 252
#> [1279] 252 252 252 252 252 252 252 252 252 252 252 252 252 252 252 252 252 252
#> [1297] 252 252 252 252 252 252 252 252 252 252 252 252 252 252 252 252 252 252
#> [1315] 252 252 252 252 252 252 252 252 252 252 252 252 252 252 252 252 252 252
#> [1333] 252 252 252 252 252 252 252 252 252 252 252 252 252 252 252 252 252 252
#> [1351] 252 252 252 252 252 252 252 252 252 252 252 252 252 252 252 252 252 252
#> [1369] 252 252 252 252 252 252 252 252 252 252 252 252 252 252 252 252 252 252
#> [1387] 252 252 252 252 252 252 252 252 252 252 252 252 252 252 252 252 252 252
#> [1405] 252 252 252 252 253 253 253 253 253 253 253 253 253 253 253 253 253 253
#> [1423] 253 253 253 253 253 253 253 253 253 253 253 253 253 253 253 253 253 253
#> [1441] 253 253 253 253 253 253 253 253 253 253 253 253 253 253 253 253 253 253
#> [1459] 253 253 253 253 253 253 253 253 253 253 253 253 253 253 253 253 253 253
#> [1477] 253 253 253 253 253 253 253 253 253 253 253 253 253 253 253 253 253 253
#> [1495] 253 253 253 253 253 253 253 253 253 253 253 253 253 253 253 253 253 253
#> [1513] 253 253 253 253 253 253 253 253 253 253 253 253 253 253 253 253 253 253
#> [1531] 253 253 253 253 253 253 253 253 253 253 253 253 253 253 253 253 253 253
#> [1549] 253 253 253 253 253 253 253 253 253 253 253 253 253 253 253 253 253 253
#> [1567] 253 253 253 253 253 253 253 253 253 253 253 253 253 253 253 253 253 253
#> [1585] 253 253 253 253 253 253 253 253 253 253 253 253 253 253 253 253 253 253
#> [1603] 253 253 253 253 253 253 253 253 253 253 253 253 253 253 253 253 253 253
#> [1621] 253 253 253 253 253 253 253 253 253 253 253 253 253 253 253 253 253 253
#> [1639] 253 253 253 253 253 253 253 253 253 253 253 253 253 253 253 253 253 253
#> [1657] 253 253 253 253 253 253 253 253 253 253 253 253 253 253 253 253 253 253
#> [1675] 253 253 253 253 253 253 253 253 253 253 254 254 254 254 254 254 254 254
#> [1693] 254 254 254 254 254 254 254 254 254 254 254 254 254 254 254 254 254 254
#> [1711] 254 254 254 254 254 254 254 254 254 254 254 254 254 254 254 254 254 254
#> [1729] 254 254 254 254 254 254 254 254 254 254 254 254 254 254 254 254 254 254
#> [1747] 254 254 254 254 254 254 254 254 254 254 254 254 254 254 254 254 254 254
#> [1765] 254 254 254 254 254 254 254 254 254 254 254 254 254 254 254 254 254 254
#> [1783] 254 254 254 254 254 254 254 254 254 254 254 254 254 254 254 254 254 254
#> [1801] 254 254 254 254 254 254 254 254 254 254 254 254 254 254 254 254 254 254
#> [1819] 254 254 254 254 254 254 254 254 254 254 254 254 254 254 254 254 254 254
#> [1837] 254 254 254 254 254 254 254 254 254 254 254 254 254 254 254 254 254 254
#> [1855] 254 254 254 254 254 254 254 254 254 254 254 254 254 254 254 254 254 254
#> [1873] 254 254 254 254 254 254 254 254 254 254 254 254 254 254 254 254 254 254
#> [1891] 254 254 254 254 254 254 254 254 254 254 254 254 254 254 254 254 254 254
#> [1909] 254 254 254 254 254 254 254 254 254 254 254 254 254 254 254 254 254 254
#> [1927] 254 254 254 254 254 254 254 254