Convert a NEXRAD polar volume file to an ODIM polar volume file

Convert a NEXRAD polar volume file to an ODIM polar volume file

nexrad_to_odim(
  pvolfile_nexrad,
  pvolfile_odim,
  verbose = FALSE,
  mount = dirname(pvolfile_nexrad),
  local_install
)

Arguments

pvolfile_nexrad
Polar volume input file in RSL format.

pvolfile_odim
Filename for the polar volume in ODIM HDF5 format to be generated.

verbose
logical. When TRUE, pipe Docker stdout to R console. On Windows always TRUE.

mount
character. String with the mount point (a directory path) for the Docker container.

local_install
character. String with path to local vol2bird installation (e.g. "/your/vol2bird_install_directory/vol2bird/bin/vol2bird"). To use local installation instead of Docker container, see details.

Value

TRUE on success

Examples

if (FALSE) {
# download a NEXRAD file, save as KBGM_example
download.file(paste("https://noaa-nexrad-level2.s3.amazonaws.com/",
  "2019/10/01/KBGM/KBGM20191001_000542_V06",
  sep = ""
), "~/KBGM_example")

# convert to ODIM format
nexrad_to_odim("~/KBGM_example", "~/KBGM_example.h5")

# verify that we have generated a polar volume in ODIM HDF5 format
get_odim_object_type("~/KBGM_example.h5")

# clean up
file.remove("~/KBGM_example", "~/KBGM_example.h5")
}