Regularize a time series of vertical profiles (vpts) on a regular time grid

Projects objects of class vpts on a regular time grid

regularize_vpts(ts, interval = "auto", t.min = ts$daterange[1],
  t.max = ts$daterange[2], units = "mins", fill = FALSE, verbose = TRUE)

Arguments

ts
An object inhereting from class vpts, see vpts for details.

interval
Time interval grid to project on. When 'auto' the median interval in the time series is used.

t.min
Start time of the projected time series, as a POSIXct object. Taken from ts when 'auto'.

t.max
End time of the projected time series, as a POSIXct object. Taken from ts when 'auto'.

units
Optional units of interval, one of 'secs', 'mins', 'hours','days', 'weeks'. Defaults to 'mins'.

fill
Logical, whether to fill missing timesteps with the values of the closest neighbouring profile.

verbose
Logical, when TRUE prints text to console.

Value

An object of class vpts with regular time steps.

Details

Irregular time series of profiles are typically aligned on a regular time grid with the expected time interval at which a radar provides data. Empty profiles with only missing data values will be inserted at time stamps of the regular time grid that have no matching profile in the irregular time series. This also has the benefit that missing profiles become visible in profile plots of regular time series using plot.vpts.

Examples

# locate example file:
vptsfile <- system.file("extdata", "vpts.txt", package = "bioRad")
# load time series:
ts <- read_vpts(vptsfile, radar = "KBGM", wavelength = "S")
# regularize the time series on a 5 minute interval grid
tsRegular <- regularize_vpts(ts, interval = 5)