Class pvol: a polar volume

Class pvol for a polar volume, and its associated R base functions.

# S3 method for pvol
summary(object, ...)

is.pvol(x)

Arguments

object
Object of class pvol.

...
Additional arguments affecting the summary produced.

x
Object of class pvol.

Value

for is.pvol: TRUE if its argument is of class pvol

Details

An object of class pvol is a list containing:

radar

character string with the radar identifier

datetime

nominal time of the volume [UTC]

scans

a list with scan objects of class 'scan'

attributes

list with the volume's \what, \where and \how attributes

geo

geographic data, a list with:

lat

latitude of the radar [decimal degrees]

lon

longitude of the radar [decimal degrees]

height

height of the radar antenna [metres above sea level]

Examples

# locate example volume file:
pvol <- system.file("extdata", "volume.h5", package = "bioRad")
# print the local path of the volume file:
pvol

#> [1] "/Users/amd427/git/bioRad/inst/extdata/volume.h5"

# load the file:
vol <- read_pvolfile(pvol)
# print summary info for the loaded polar volume:
vol

#>                Polar volume (class pvol)
#> 
#>      # scans:  3 
#>        radar:  SE50 
#>       source:  WMO:02606,RAD:SE50,PLC:Angelholm,NOD:seang,ORG:82,CTY:643,CMT:Swedish radar 
#> nominal time:  2015-10-18 18:00:00 
#> 

# print summary info for the scans in the polar volume:
vol$scans

#> [[1]]
#>                   Polar scan (class scan)
#> 
#>      parameters:  DBZH VRADH RHOHV ZDR PHIDP 
#> elevation angle:  0.5 deg
#>            dims:  480 bins x 360 rays
#> 
#> [[2]]
#>                   Polar scan (class scan)
#> 
#>      parameters:  DBZH VRADH RHOHV ZDR PHIDP 
#> elevation angle:  1.5 deg
#>            dims:  480 bins x 360 rays
#> 
#> [[3]]
#>                   Polar scan (class scan)
#> 
#>      parameters:  DBZH VRADH RHOHV ZDR PHIDP 
#> elevation angle:  2.5 deg
#>            dims:  480 bins x 360 rays
#> 

# copy the first scan to a new object 'scan'
scan <- vol$scans[[1]]
is.pvol("this is not a polar volume but a string") # > FALSE

#> [1] FALSE