Class vp: a vertical profile of animals

Class vp for a vertical profile of animals

# S3 method for vp
summary(object, ...)

is.vp(x)

# S3 method for vp
dim(x)

Arguments

object
An object of class vp.

...
Additional arguments affecting the summary produced.

x
An object of class vp.

Value

For is.vp: TRUE if its argument is of class vp.

For dim.vp: dimensions of the profile data.

Details

An object of class vp contains a vertical profile. A vertical profile contains a collection of quantities, with each quantity having values at different altitude layers above the earth's surface, typically equally spaced altitudinal layers.

Data contained in this class object should be accessed with the get_quantity function. Information stored under attributes (see below) can be accessed directly.

A vp object is a list containing

radar

the radar identifier

datetime

the nominal time of the profile

data

the profile data, a list containing:
height

height above mean sea level (m). Alt. bin from height to height+interval)

u

speed component west to east (m/s)

v

speed component north to south (m/s)

w

vertical speed (unreliable!) (m/s)

ff

horizontal speed (m/s)

dd

direction (degrees, clockwise from north)

sd_vvp

VVP radial velocity standard deviation (m/s)

gap

Angular data gap detected (T/F)

dbz

Animal reflectivity factor (dBZ)

eta

Animal reflectivity (cm^2/km^3)

dens

Animal density (animals/km^3)

DBZH

Total reflectivity factor (bio+meteo scattering) (dBZ)

n

number of points VVPvelocity analysis (u,v,w,ff,dd)

n_all

number of points VVP st.dev. estimate (sd_vvp)

n_dbz

number of points density estimate (dbz,eta,dens)

n_dbz_all

number of points total reflectivity estimate (DBZH)

attributes

list with the profile's \what, \where and \how attributes

Conventions

NA Maps to 'nodata' in the ODIM convention: value to denote areas void of data (never radiated)

NaN Maps to 'undetect' in the ODIM convention: denote areas below the measurement detection threshold (radiated but nothing detected). The value is also used when there are too few datapoints to calculate a quantity.

0 Maps to 0 in the ODIM convention: denote areas where the quantity has a measured value of zero (radiated and value zero detected or inferred).

It depends on a radar's detection threshold or signal to noise ratio whether it safe to assume an 'undetect' is equivalent to zero. When dealing with close range data only (within 35 km), it is typically safe to assume aerial densities (dens) and reflectivities (eta) are in fact zero in case of undetects.

Examples

# load example vp object
data(example_vp)
example_vp

#>                Vertical profile (class vp)
#> 
#>        radar:  seang 
#>       source:  WMO:02606,RAD:SE50,PLC:Angelholm,NOD:seang,ORG:82,CTY:643,CMT:Swedish radar 
#> nominal time:  2015-10-18 18:00:00 
#> generated by:  vol2bird 0.3.17 

# check that the object is a vp object:
is.vp(example_vp)

#> [1] TRUE

# dimensions of the vp object:
dim(example_vp)

#> [1] 25 16