Inspect a vertical profile (vp)

R base functions for inspecting a vertical profile of biological targets (vp) object.

# S3 method for vp
summary(object, ...)

is.vp(x)

# S3 method for vp
dim(x)

Arguments

object
A vp object.

...
Additional arguments affecting the summary produced.

x
A vp object.

Value

For is.vp(): TRUE for an object of class vp, otherwise FALSE.

For dim.vp(): number of heights and quantities in a vertical profile (vp).

Details

A vertical profile of biological targets contains a collection of quantities, organized in different (typically equally spaced) altitude layers (height bins) above the earth's surface. A vertical profile (vp) object is a list containing:

radar: Radar identifier.

datetime: Nominal time of the volume to which the scan belongs in UTC.

data: A data.frame with the profile's quantities organized per height bin. Use get_quantity() to access these:

height: Height bin (lower bound) in m above sea level.

u: Speed component west to east in m/s.

v: Speed component south to north in m/s.

w: Vertical speed (unreliable!) in m/s.

ff: Horizontal speed in m/s.

dd: Direction in degrees clockwise from north.

sd_vvp: VVP radial velocity standard deviation in m/s.

gap: Angular data gap detected in T/F.

dbz: Animal reflectivity factor in dBZ.

eta: Animal reflectivity in cm^2/km^3.

dens: Animal density in animals/km^3.

DBZH: Total reflectivity factor (bio + meteo scattering) in dBZ.

n: Number of data points used for the ground speed estimates (quantities u, v, w, ff, dd).

n_all: Number of data points used for the radial velocity standard deviation estimate (quantity sd_vvp).

n_dbz: Number of data points used for reflectivity-based estimates (quantities dbz, eta, dens).

n_dbz_all: Number of data points used for the total reflectivity estimate (quantity DBZH).

attributes: List of the vertical profile's what, where and how attributes.

Conventions

NA: Maps to nodata in the ODIM convention: value to denote areas void of data (never radiated).

NaN: Maps to undetect in the ODIM convention: denote areas below the measurement detection threshold (radiated but nothing detected). The value is also used when there are too few datapoints to calculate a quantity.

0: Maps to 0 in the ODIM convention: denote areas where the quantity has a measured value of zero (radiated and value zero detected or inferred).

It depends on a radar's detection threshold or signal to noise ratio whether it safe to assume an undetect is equivalent to zero. When dealing with close range data only (within 35 km), it is typically safe to assume aerial densities (dens) and reflectivities (eta) are in fact zero in case of undetects.

See also

Examples

# Verify that an object is of class scan
is.vp(example_vp)

#> [1] TRUE

# Get summary info
example_vp # Same as summary(example_vp) or print(example_vp)

#>                Vertical profile (class vp)
#> 
#>        radar:  seang 
#>       source:  WMO:02606,RAD:SE50,PLC:Angelholm,NOD:seang,ORG:82,CTY:643,CMT:Swedish radar 
#> nominal time:  2015-10-18 18:00:00 
#> generated by:  vol2bird 0.3.17 

# Get dimensions
dim(example_vp)

#> [1] 25 16