Calculate sunrise or sunset for a time and place

Calculate sunrise or sunset for a time and place

sunrise(date, lon, lat, elev = -0.268, tz = "UTC")

sunset(date, lon, lat, elev = -0.268, tz = "UTC")

Arguments

date
Date inheriting from class POSIXt or a string interpretable by as.Date.

lon
Longitude in decimal degrees.

lat
Latitude in decimal degrees.

elev
Sun elevation in degrees.

tz
output time zone. Ignored if date has an associated time zone already

Value

The moment of sunrise or sunset in UTC time.

Details

The angular diameter of the sun is about 0.536 degrees, therefore the moment of sunrise/sunset corresponds to half that elevation at -0.268 degrees.

This is a convenience function mapping to crepuscule.

Approximate astronomical formula are used, therefore the moment of sunrise / sunset may be off by a few minutes

Examples

# sunrise in the Netherlands
sunrise("2016-01-01", 5, 53)

#> [1] "2016-01-01 07:58:17 UTC"

# sunset in the Netherlands
sunset("2016-01-01", 5, 53)

#> [1] "2016-01-01 15:28:31 UTC"

# civil twilight in Ithaca, NY, today
sunrise(Sys.time(), -76.5, 42.4, elev = -6)

#> [1] "2019-07-31 04:25:14 CDT"